@@ -15,7 +15,7 @@
         "issues": "http://github.com/Innmind/Ark/issues"
     },
     "require": {
-        "php": "~7.2",
+        "php": "~7.4",
         "innmind/url": "^2.0",
         "innmind/immutable": "^2.8",
         "ovh/ovh": "^2.0",
